requirements.txt
pytest==8.1.1
requests==2.31.0
msal==1.28.0
python-dotenv==1.0.1
pyjwt==2.8.0